# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WorkPilot API",
    "summary": "Evidence-grounded work summary agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_api_v1_tasks_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRecord"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/events": {
      "get": {
        "summary": "Task Events",
        "operationId": "task_events_api_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/materials": {
      "get": {
        "summary": "Materials",
        "operationId": "materials_api_v1_tasks__task_id__materials_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportedPackage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/facts": {
      "get": {
        "summary": "Facts",
        "operationId": "facts_api_v1_tasks__task_id__facts_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/draft": {
      "get": {
        "summary": "Draft",
        "operationId": "draft_api_v1_tasks__task_id__draft_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/decisions": {
      "post": {
        "summary": "Decide",
        "operationId": "decide_api_v1_tasks__task_id__decisions_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_api_v1_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_v1_tasks__task_id__history_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response History Api V1 Tasks  Task Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/usage": {
      "get": {
        "summary": "Usage",
        "operationId": "usage_api_v1_tasks__task_id__usage_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Usage Api V1 Tasks  Task Id  Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/artifacts/{name}": {
      "get": {
        "summary": "Artifact",
        "operationId": "artifact_api_v1_tasks__task_id__artifacts__name__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations": {
      "post": {
        "summary": "Create Evaluation",
        "operationId": "create_evaluation_api_v1_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Evaluation Api V1 Evaluations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "summary": "Get Evaluation",
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Evaluation Api V1 Evaluations  Evaluation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_task_api_v1_tasks_post": {
        "properties": {
          "report_type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "period": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Period"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "period",
          "files"
        ],
        "title": "Body_create_task_api_v1_tasks_post"
      },
...
      "DecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "edit",
              "reject"
            ],
            "title": "Decision"
          },
          "edited_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Markdown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionRequest"
      },
      "EvaluationRequest": {
        "properties": {
          "split": {
            "type": "string",
            "enum": [
              "dev",
              "test",
              "all"
            ],
            "title": "Split"
          },
          "system": {
            "type": "string",
            "enum": [
              "baseline",
              "workpilot_single",
              "workpilot_heterogeneous",
              "challenger",
              "both",
              "all_systems",
              "dev_matrix",
              "test_matrix"
            ],
            "title": "System"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "split",
          "system"
        ],
        "title": "EvaluationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportedPackage": {
        "properties": {
          "materials": {
            "items": {
              "$ref": "#/components/schemas/Material"
            },
            "type": "array",
            "minItems": 1,
            "title": "Materials"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "materials",
          "evidence"
        ],
        "title": "ImportedPackage"
      },
      "Material": {
        "properties": {
          "material_id": {
            "type": "string",
            "pattern": "^mat_[0-9a-f]{16}$",
            "title": "Material Id"
          },
          "source_path": {
            "type": "string",
            "minLength": 1,
            "title": "Source Path"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "markdown",
              "text",
              "csv"
            ],
            "title": "Media Type"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "material_id",
          "source_path",
          "media_type",
          "sha256",
          "size_bytes"
        ],
        "title": "Material"
      },
      "ReportBundle": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/ReportMetadata"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Facts"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/Risk"
            },
            "type": "array",
            "title": "Risks"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSection"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sections"
          },
          "markdown": {
            "type": "string",
            "minLength": 1,
            "title": "Markdown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metadata",
          "facts",
          "sections",
          "markdown"
        ],
        "title": "ReportBundle"
      },
...
      "ReportType": {
        "type": "string",
        "enum": [
          "work_summary",
          "performance_review"
        ],
        "title": "ReportType"
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "ingesting",
          "extracting",
          "validating",
          "verifying",
          "drafting",
          "waiting_approval",
          "completed",
          "rejected",
          "failed",
          "recovering",
          "cancelled"
        ],
        "title": "RunStatus"
      },
      "TaskRecord": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "report_type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "report_type",
          "period",
          "status",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRecord"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationResult": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Facts"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/Risk"
            },
            "type": "array",
            "title": "Risks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VerificationResult"
      },
      "WorkflowStage": {
        "type": "string",
        "enum": [
          "queued",
          "ingesting",
          "extracting",
          "validating",
          "verifying",
          "drafting",
          "awaiting_approval",
          "completed",
          "rejected",
          "failed",
          "recovering",
          "cancelled"
        ],
        "title": "WorkflowStage"
      }
    }
  }
}